refactor: split prepare.ts
