+++ examples/contracts.build.ts
@@ -2,12 +2,15 @@ import fs from "fs";
 import { run } from "../src/node";
 import path from "path";
 import { glob } from "glob";
-import { verify } from "../src/verify";
+import { verify } from "./verify";
 import { Logger } from "../src/context/logger";
 import { __DANGER__disableVersionNumber } from "../src/pipeline/version";
 
+const configPath = path.join(__dirname, "tact.config.json");
+const packagesPath = path.resolve(__dirname, "output", "*.pkg");
+
 // Read cases
-void (async () => {
+const main = async () => {
     // Disable version number in packages
     __DANGER__disableVersionNumber();
 
@@ -16,18 +19,14 @@ void (async () => {
     try {
         // Compile projects
         const compileResult = await run({
-            configPath: path.join(__dirname, "..", "tact.config.json"),
+            configPath,
         });
         if (!compileResult.ok) {
             throw new Error("Tact projects compilation failed");
         }
 
         // Verify projects
-        for (const pkgPath of glob.sync(
-            path.normalize(
-                path.resolve(__dirname, "..", "examples", "output", "*.pkg"),
-            ),
-        )) {
+        for (const pkgPath of glob.sync(path.normalize(packagesPath))) {
             const res = await verify({
                 pkg: fs.readFileSync(pkgPath, "utf-8"),
             });
@@ -39,4 +38,6 @@ void (async () => {
         logger.error(error as Error);
         process.exit(1);
     }
-})();
+};
+
+void main();
