+++ examples/verify.ts
@@ -1,10 +1,10 @@
 import normalize from "path-normalize";
 import { Cell } from "@ton/core";
-import { Config, Options } from "./config/parseConfig";
-import { ILogger, Logger } from "./context/logger";
-import { PackageFileFormat, run } from "./";
-import { fileFormat } from "./packaging/fileFormat";
-import { getCompilerVersion } from "./pipeline/version";
+import { Config, Options } from "../src/config/parseConfig";
+import { ILogger, Logger } from "../src/context/logger";
+import { PackageFileFormat, run } from "../src";
+import { fileFormat } from "../src/packaging/fileFormat";
+import { getCompilerVersion } from "../src/pipeline/version";
 
 export type VerifyResult =
     | {
