+++ knip.json
@@ -4,6 +4,7 @@
   "project": ["src/**/*.ts", "bin/tact.js", "bin/unboc.js"],
   "ignore": [
     "src/ast/ast.ts",
+    "src/config/parseConfig.ts",
     "src/prettyPrinter.ts",
     "src/error/display-to-json.ts",
     "src/grammar/src-info.ts",
