+++ src/index.ts
@@ -31,6 +31,5 @@ export { VirtualFileSystem } from "./vfs/VirtualFileSystem";
 export { createVirtualFileSystem } from "./vfs/createVirtualFileSystem";
 
 export * from "./browser";
-export * from "./verify";
 export * from "./context/logger";
 export * from "./error/errors";
