+++ src/test/benchmarks/contracts/functions.tact
